3d camera

+++ ConsoleApplication2/GameObject.cpp
@@ -5,5 +5,5 @@
 GameObject::GameObject(){}
 GameObject::~GameObject(){}
 void GameObject::update(double delta_t){}
-void GameObject::setSize(int x, int y, int z){ _size.set(x, y, z); }
+void GameObject::setSize(double x, double y, double z){ _size.set(x, y, z); }
 Vector3 GameObject::getSize(){ return _size; }
