+++ ConsoleApplication2/GameObject.h
@@ -12,7 +12,7 @@ class GameObject : public Entity
 		~GameObject();
 		virtual void draw() = 0;
 		virtual void update(double delta_t);
-		void setSize(int x, int y, int z);
+		void setSize(double x, double y, double z);
 		Vector3 getSize();
 };
